Change DB to Postgres
Uses knex and pg to connect and build queries. Migrations use knex, found in the top-level "migrations" directory.

Removes the Mongoose and Typegoose dependencies.

Tests are just skipped for now, since they're trivial anyway, and it's not worth fixing them until Apollo Server is implemented.

When tests are reimplemented, consider using a real Postgres connection and seed data for integration tests to maximize their similarity to real world. In-memory DBs (with SQLite) could work, but the SQLite and PostgreSQL dialects are somewhat different, and that is likely to cause issues.